# docs/swagger.yaml
definitions:
  errorcode.Error:
    type: object
...
  requests.KubeApplyYamlClusterRequest:
    properties:
      name:
        type: string
      yaml:
        type: string
    type: object
...
  requests.KubeNamespaceCreateRequest:
    properties:
      annotations:
        type: object
      description:
        description: 描述信息（可选） -> 写入 annotation
        type: string
      labels:
        description: Labels（可选）
        type: object
      name:
        description: Namespace 名称（必填）
        type: string
      quota_cpu:
        type: string
      quota_memory:
        type: string
      quota_pods:
        type: string
    type: object
  requests.KubeNamespaceLabelPatchRequest:
    properties:
      add:
        additionalProperties:
          type: string
        description: 添加/更新的标签
        type: object
      name:
        type: string
      remove:
        description: 要删除的标签 key
        items:
          type: string
        type: array
    type: object
  requests.KubeNamespaceUpdateRequest:
    type: object
...
  response.Response:
    type: object
info:
  contact: {}
  description: 基于Gin+Vue开发的K8s管理平台
  termsOfService: https://gitee.com/jay-kim/k8s_operation
  title: K8s管理平台
  version: "1.0"
paths:
...
  /api/v1/k8s/namespace/apply_yaml:
    put:
      consumes:
      - application/json
      description: 应用修改后的 YAML 配置到 Namespace
      parameters:
      - description: YAML内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeApplyYamlClusterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 应用 Namespace YAML 配置
      tags:
      - K8s Namespace 管理
  /api/v1/k8s/namespace/create:
    post:
      consumes:
      - application/json
      description: 创建命名空间，并可选设置 labels/annotations 和资源配额（CPU/Memory/Pods）
      parameters:
      - description: Namespace 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeNamespaceCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 创建 Namespace
      tags:
      - K8s Namespace 管理
  /api/v1/k8s/namespace/delete:
    delete:
      description: 删除指定 Namespace（级联删除内部资源）
      parameters:
      - description: Namespace 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 删除 Namespace
      tags:
      - K8s Namespace 管理
  /api/v1/k8s/namespace/detail:
    get:
      description: 查询指定 Namespace 的详细信息，包含资源统计
      parameters:
      - description: Namespace 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 Namespace 详情
      tags:
      - K8s Namespace 管理
  /api/v1/k8s/namespace/labels:
    patch:
      consumes:
      - application/json
      description: 添加或删除 Namespace 标签
      parameters:
      - description: 标签修改参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.KubeNamespaceLabelPatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 修改 Namespace 标签
      tags:
      - K8s Namespace 管理
  /api/v1/k8s/namespace/list:
    get:
      description: 支持分页、名称模糊查询，返回每个命名空间的资源统计
      parameters:
      - description: 命名空间名称(模糊匹配)
        in: query
        maxLength: 100
        name: name
        type: string
      - description: 页码(从1开始)
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量(默认20)
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 Namespace 列表
      tags:
      - K8s Namespace 管理
  /api/v1/k8s/namespace/patch:
    patch:
      description: 支持新增、更新、删除 labels 与 annotations
      parameters:
      - description: Namespace 名称
        in: query
        name: name
        required: true
        type: string
      - description: Patch 内容
        in: body
        name: patch
        required: true
        schema:
          $ref: '#/definitions/requests.KubeNamespaceUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 修改 Namespace（labels / annotations）
      tags:
      - K8s Namespace 管理
  /api/v1/k8s/namespace/yaml:
    get:
      description: 获取指定 Namespace 的 YAML 配置
      parameters:
      - description: Namespace 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 Namespace YAML 配置
      tags:
      - K8s Namespace 管理
...
swagger: "2.0"